requirements.txt
django ~= 3.1
aiohttp ~= 3.7
daphne ~= 3.0
twisted[tls, http2]
google-auth
requests
drf_yasg == 1.20.0
whitenoise == 5.2.0
djangorestframework ~= 3.12
